fix: :bug: 修复一些实体&地图的漏洞
①实体现在能正确被删除②地图现在正确拥有「有限无界」特性

+++ src/batr/game/mods/native/tool/Weapon.ts
@@ -108,7 +108,7 @@ export default class Weapon extends Tool {
 		baseDamage: uint = 1,
 		reverseCharge: boolean = false
 	) {
-		super(id, baseCD_S * FIXED_TPS, chargeMaxTime_S * FIXED_TPS);
+		super(id, uint(baseCD_S * FIXED_TPS), uint(chargeMaxTime_S * FIXED_TPS));
 		// defaultCD,defaultChargeTime instanceof Per Second
 		this._baseDamage = baseDamage;
 		this._reverseCharge = reverseCharge;
